requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.24.0
asyncpg==0.29.0
bcrypt==4.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
pydantic==2.11.7
httpx==0.27.0
psycopg2-binary==2.9.9
